+++ src/Pim/Bundle/CatalogBundle/Entity/Media.php
@@ -2,7 +2,6 @@
 
 namespace Pim\Bundle\CatalogBundle\Entity;
 
-use Doctrine\ORM\Mapping as ORM;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Symfony\Component\HttpFoundation\File\File;
 
@@ -12,18 +11,11 @@ use Symfony\Component\HttpFoundation\File\File;
  * @author    Gildas Quemener <gildas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @ORM\Table(name="pim_catalog_media")
- * @ORM\Entity
  */
 class Media
 {
     /**
      * @var integer $id
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
      */
     protected $id;
 
@@ -38,8 +30,6 @@ class Media
      * Filename
      *
      * @var string $filename
-     *
-     * @ORM\Column(name="filename", type="string", length=255, unique=true, nullable=true)
      */
     protected $filename;
 
@@ -47,8 +37,6 @@ class Media
      * File path
      *
      * @var string $filePath
-     *
-     * @ORM\Column(name="filepath", type="string", length=255, unique=true, nullable=true)
      */
     protected $filePath;
 
@@ -56,8 +44,6 @@ class Media
      * Original file name
      *
      * @var string $originalFilename
-     *
-     * @ORM\Column(nullable=true)
      */
     protected $originalFilename;
 
@@ -65,18 +51,9 @@ class Media
      * Mime type
      *
      * @var string $mimeType
-     *
-     * @ORM\Column(name="mimeType", type="string", length=255, nullable=true)
      */
     protected $mimeType;
 
-    /**
-     * @ORM\OneToOne(
-     *     targetEntity="Pim\Bundle\CatalogBundle\Model\ProductValueInterface",
-     *     mappedBy="media"
-     * )
-     * @ORM\JoinColumn(name="value_id", referencedColumnName="id", onDelete="SET NULL")
-     */
     protected $value;
 
     protected $removed = false;