+++ src/Pim/Bundle/ImportExportBundle/Normalizer/FlatCategoryNormalizer.php
@@ -2,8 +2,6 @@
 
 namespace Pim\Bundle\ImportExportBundle\Normalizer;
 
-use Pim\Bundle\CatalogBundle\Model\CategoryInterface;
-
 /**
  * A normalizer to transform a category entity into a flat array
  *