+++ CHANGELOG.md
@@ -11,10 +11,12 @@
 - Group types management
 - Import/Export product groups (CSV)
 - Import/Export associations (CSV)
+- Import/Export product associations (CSV)
 - Import/Export attributes (CSV)
 - Import/Export attribute options (CSV)
-- Upload and import an archive
+- Upload and import an archive (CSV and medias)
 - Download an archive containing the exported products along with media
+- Add the column "enabled" in the CSV file for products import/export and for versioning
 
 ## Improvements
 - Export media into separated sub directories
@@ -24,10 +26,12 @@
 - Display the "has product" filter by default in the product grid of group edit view
 - Display currency label in currencies datagrid
 - Disable changing the code of all configuration-related entities
+- Merge the directory and filename of export profiles into a single file path property
 
 ## Bug fixes
 - Mass delete products
 - Fix some issues with import ACL translations (issues#484)
+- Add a message when trying to delete an attribute used by one or more variant groups instead of throwing an error
 
 ## BC breaks
 - Command "pim:product:completeness-calculator" has been replaced into "pim:completeness:calculate"